Multiple ammends, see description.
Added Doxygen and config files as well as 2 batch scripts to generate. gen_docs generates and opens the index.html whereas regen_docs only generates.
The singleton renderer in the Game class and the Log singleton have been readded and so various files are changing the calls to GetRenderer() or Log::Get().
There is also a change to entity which does not ask the texture cache for a texture if the file path is empty. This was causing a crash in the example project.