Rectangle attempt didnt work. Unparented UI_Panel Changes to Texture render function, added todo to trello to fix other functions. Added multiple constructors to WorldRectangle to make it easier to use in future. Also added a render function to visualise these rectangles. Added new test image, lil stick fella :)
SpatialHash retrieve now has actual code content! It generates the id for the corners of the WorldRect and loops between them, collecting the IDs, then for each ID it dumps the entities from the map into a vector. It then checks their position is actually within the rect and if not they are removed. The vector is then returned to the user full of entities.
Minor changes to HashBucket - exposing the data for the hashmap to use as well as renaming the buckets variable.
World has new functions to allow users to get a list of entities within a WorldRect, this will be used by the editor in the future.