Filled out the script loader with the functions from the Scripting project.
hTech.vcxproj - Changes because we added files.
World.cpp - Added lines to create some entities in the world that we can use to test the scripts out.
       The two png's that were added were so we can visualise the differences between the two objects with different scripts.
Scripting/CPPScripting.cpp - Removed unused code, set a string value and added some error checks.

hTech/ScriptLoader - ScriptLoader is a class of static functions which are called to get ScriptObjects by ScriptComponents. It handles all the loading of function pointers and stores them in a map to be reused to avoid loading them multiple times. The DLL handle is also maintained and handled in this to avoid it being accessed by any other functions.
hTech/ScriptObject - This is the class which the custom scripts inherit from, similar to MonoBehaviour, containing all the function definitions that can be overridden.
hTech/Component_Script - The actual script component. Contains a ScriptObject variable which is instantiated in the constructor via the ScriptLoader.