Player mode.
Moved a large amount of code from Engine.cpp into Editor.cpp so that the editor is not rendered in Player mode. Added support for command line loading of projects too so that the argv[1] can be a project file location and will load automatically.
Fixed a stack overflow bug caused by repeatedly loading the library when getting script functions.